Re-structure cairo writer
