+++ src/cairoWriter/base.ts
@@ -1,0 +1,21 @@
+import { AST } from '../ast/ast';
+import { ASTNodeWriter } from 'solc-typed-ast';
+import { NotSupportedYetError } from '../utils/errors';
+
+export abstract class CairoASTNodeWriter extends ASTNodeWriter {
+  ast: AST;
+  throwOnUnimplemented: boolean;
+  constructor(ast: AST, throwOnUnimplemented: boolean) {
+    super();
+    this.ast = ast;
+    this.throwOnUnimplemented = throwOnUnimplemented;
+  }
+
+  logNotImplemented(message: string) {
+    if (this.throwOnUnimplemented) {
+      throw new NotSupportedYetError(message);
+    } else {
+      console.log(message);
+    }
+  }
+}
