+++ src/cairoWriter/writers/sourceUnitWriter.ts
@@ -1,0 +1,80 @@
+import assert from 'assert';
+import { ASTWriter, ContractKind, SourceUnit, SrcDesc } from 'solc-typed-ast';
+import { getStructsAndRemappings, removeExcessNewlines } from '../../export';
+import { CairoASTNodeWriter } from '../base';
+import { writeImports } from '../utils';
+
+// Used by CairoContractWriter
+export const interfaceNameMappings: Map<SourceUnit, Map<string, string>> = new Map();
+// Used by IdentifierWriter
+export let structRemappings: Map<number, string>;
+
+export class SourceUnitWriter extends CairoASTNodeWriter {
+  writeInner(node: SourceUnit, writer: ASTWriter): SrcDesc {
+    this.generateInterfaceNameMappings(node);
+
+    // Every sourceUnit should only define a single contract
+    const mainContract_ = node.vContracts.filter((cd) => cd.kind !== ContractKind.Interface);
+    assert(mainContract_.length <= 1, 'There should only be one active contract per sourceUnit');
+    const [mainContract] = mainContract_;
+
+    const [freeStructs, freeStructRemappings_] = mainContract
+      ? getStructsAndRemappings(node, this.ast)
+      : [[], new Map()];
+    structRemappings = freeStructRemappings_;
+
+    // Only constants generated by `newToDeploy` exist at this stage
+    const constants = node.vVariables.flatMap((v) => {
+      assert(v.vValue !== undefined, 'Constants cannot be unanssigned');
+      return [`// ${v.documentation}`, `const ${v.name} = ${writer.write(v.vValue)};`].join('\n');
+    });
+
+    const structs = [...freeStructs, ...node.vStructs, ...(mainContract?.vStructs || [])].map((v) =>
+      writer.write(v),
+    );
+
+    const functions = node.vFunctions.map((v) => writer.write(v));
+
+    const contracts = node.vContracts.map((v) => writer.write(v));
+
+    const generatedUtilFunctions = this.ast.getUtilFuncGen(node).getGeneratedCode();
+    const imports = writeImports(this.ast.getImports(node));
+    return [
+      removeExcessNewlines(
+        [
+          '%lang starknet',
+          [imports],
+          ...constants,
+          ...structs,
+          generatedUtilFunctions,
+          ...functions,
+          ...contracts,
+        ].join('\n\n\n'),
+        3,
+      ),
+    ];
+  }
+
+  private generateInterfaceNameMappings(node: SourceUnit) {
+    const map: Map<string, string> = new Map();
+    const existingNames = node.vContracts
+      .filter((c) => c.kind !== ContractKind.Interface)
+      .map((c) => c.name);
+
+    node.vContracts
+      .filter((c) => c.kind === ContractKind.Interface)
+      .forEach((c) => {
+        const baseName = c.name.replace('@interface', '');
+        const interfaceName = `${baseName}_warped_interface`;
+        if (!existingNames.includes(baseName)) {
+          map.set(baseName, interfaceName);
+        } else {
+          let i = 1;
+          while (existingNames.includes(`${interfaceName}_${i}`)) ++i;
+          map.set(baseName, `${interfaceName}_${i}`);
+        }
+      });
+
+    interfaceNameMappings.set(node, map);
+  }
+}
